Rename isHostState file, update imports.
