+++ src/components/common/Host/Host.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useRecoilValue } from 'recoil';
 
-import { isHostState } from '../../../state/isHostAtom';
+import { isHostState } from '../../../state/isHostState';
 import { HostPlayerProps } from '../commonTypes';
 
 const Host = (props: HostPlayerProps): React.ReactElement => {
