+++ src/components/common/Player/Player.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useRecoilValue } from 'recoil';
 
-import { isHostState } from '../../../state/isHostAtom';
+import { isHostState } from '../../../state/isHostState';
 import { HostPlayerProps } from '../commonTypes';
 
 const Player = (props: HostPlayerProps): React.ReactElement => {
